fix(core): eliminate deadlock, enforce safety defaults, and strict typing
- **messaging**: Fix guaranteed deadlock in CommandBus.RegisterHandlerWithContext where the mutex was locked recursively.
- **domain**: Change entity version type from int64 to uint64 to enforce non-negative monotonic semantics.
- **server**: Enforce default HTTP timeouts (Read/Write 15s, Idle 60s) to prevent resource exhaustion.
- **validation**: Count characters by runes, not bytes, in ValidateStringLength for correct multi-byte support.
- **http**: Clarify IFileHandler contract and standardize timeout constants.

+++ domain/eventsourced/entity.go
@@ -50,8 +50,8 @@ func (a *EventSourcedAggregate[T]) GetID() T {
 }
 
 // GetVersion 实现 IEntity 接口。
-func (a *EventSourcedAggregate[T]) GetVersion() int64 {
-	return int64(a.version)
+func (a *EventSourcedAggregate[T]) GetVersion() uint64 {
+	return a.version
 }
 
 // GetAggregateType 返回聚合类型。
